Move aux builder into its own module to separate it from the general logic
