+++ src/per_test_config.rs
@@ -10,8 +10,9 @@ use std::process::{Command, Output};
 
 use spanned::Spanned;
 
+use crate::aux_builds::AuxBuilder;
 use crate::custom_flags::Flag;
-use crate::dependencies::{AuxBuilder, BuildManager};
+use crate::dependencies::BuildManager;
 pub use crate::parser::{Comments, Condition, Revisioned};
 use crate::parser::{ErrorMatch, ErrorMatchKind, OptWithLine};
 pub use crate::rustc_stderr::Level;
